Collection and Item classes
Made a few updates to the Collection and Item .java files.  Main
addition was setting an editable variable for both classes. This
variable will be used as a precondition in the set methods to inhibit a
user from editing information that is not editable.  By using
x.setEditable(True/False); you can access/lock the object x's variables
in the back end coding.